Moving `Logic/User.php` to `Logic/ConsoleOutput/Users.php`
